+++ VeriCert/src/abis/Migrations.json
@@ -1374,7 +1374,7 @@
   },
   "networks": {},
   "schemaVersion": "3.0.2",
-  "updatedAt": "2020-02-29T05:49:56.301Z",
+  "updatedAt": "2020-03-02T11:04:08.102Z",
   "devdoc": {
     "methods": {}
   },
