+++ VeriCert/src/components/Main.js
@@ -1,59 +1,10 @@
 import React, {Component} from 'react';
-import {db,auth} from '../services/firebase'
-import ResultCard from './ResultCard.js'
+import CertCreate from './CertCreate.js'
+import QueryData from './QueryData.js'
+import CertList from './CertList.js'
+
 
 class Main extends Component{
-
-  constructor(props) {
-      super(props)
-      this.state = {
-       submitted: false,
-       resultId: 0,
-       renderComplete: false
-
-      }
-
-  
-    this.queryData = this.queryData.bind(this)
-    this.renderId = this.renderId.bind(this)
-    } 
-
-
-
-    //include promise in query data, 
-  queryData(input){
-
-      db.collection('students').where("ic", "==", input)
-        .get()
-        .then(snapshot => {
-          snapshot.forEach(doc =>{
-            var id = doc.data().id           
-
-            this.setState({resultId: id} , function(){
-              this.setState({renderComplete: true})
-            } )
-         
-          })
-
-        })
-        .catch(function(error){
-          console.log(error);
-        })
-        //console.log("query function")
-    }
-
-
-
-   //query data as promise, then renderId 
-   //check if finish render, only call
-   renderId() {
-        console.log("function called")      
-        return <ResultCard resultId={this.state.resultId} certs={this.props.certs} />
-  } 
-
-
-
-
 
   render(){
     return(
@@ -64,141 +15,23 @@ class Main extends Component{
               <div className="content mr-auto ml-auto">
               <p>&nbsp;</p>
 
-              <div>
-               <form onSubmit={(event) =>{
-                event.preventDefault()
-                const ic = this.icContent.value
-                const name = this.nameContent.value
-                const course = this.courseContent.value
-                const result = this.resultContent.value
-                const date = this.dateContent.value
-                this.props.createCert(ic,name,course,result,date)
-                const id = this.props.certCount.toNumber()+1
-                this.props.addNewStudent(id,ic)
-                }}>
-
-
-                <div className="form-group mr-sm-2">
-                  <input
-                    id="icContent"
-                    type="text"
-                    ref={(input) => {this.icContent = input}}
-                    className="form-control"
-                    placeholder="Enter IC"
-                    required />
-                </div>
-
-                <div className="form-group mr-sm-2">
-                  <input
-                    id="nameContent"
-                    type="text"
-                    ref={(input) => {this.nameContent = input}}
-                    className="form-control"
-                    placeholder="Enter name"
-                    required />
-                </div>
-
-                <div className="form-group mr-sm-2">
-                  <input
-                    id="courseContent"
-                    type="text"
-                    ref={(input) => {this.courseContent = input}}
-                    className="form-control"
-                    placeholder="Enter course"
-                    required />
-                </div>
-
-                <div className="form-group mr-sm-2">
-                  <input
-                    id="resultContent"
-                    type="text"
-                    ref={(input) => {this.resultContent = input}}
-                    className="form-control"
-                    placeholder="Enter result"                  
-                    required />
-                </div>
-
-                <div className="form-group mr-sm-2">
-                  <input
-                    id="dateContent"
-                    type="text"
-                    ref={(input) => {this.dateContent = input}}                    
-                    className="form-control"
-                    placeholder="Enter issue date"                    
-                    required />
-                </div>
-
-                <button type="submit" className="btn btn-primary btn-block"> Create </button>
-              </form>
-              </div>
-
+              <CertCreate 
+                createCert={this.props.createCert}
+                addNewStudent={this.props.addNewStudent}
+                certCount={this.props.certCount}
+              />
 
               <p>&nbsp;</p>
 
-              <div>
-
-                <form onSubmit={(event) =>{
-                  event.preventDefault()
-                  const ic = this.ic.value
-                  this.queryData(ic)
-                  this.setState({ submitted: true })
-
-                }}>
-
-
-                  <label>
-                    ID:
-                    <input 
-                      id="ic"
-                      type="text"
-                      ref={(input) => {this.ic=input}}/>
-                  </label>
-
-                  <button type="submit">search</button>
-
-              </form>
-
-
-               {this.state.renderComplete && this.renderId()}
-
-              </div>
+              <QueryData
+                certs={this.props.certs} 
+              />
 
               <p>&nbsp;</p>
-              <p>&nbsp;</p>
 
-
-                  { this.props.certs.map((cert, key) => {
-                      return(
-                        <div className="card mb-4" key={key}>
-                          <div className="card-header">
-                            <small className="text-mutet"> {cert.author}</small>
-                          </div>
-
-                          <ul id="postList" className="list-group list-group-flush">
-                              <li className="list-group-item">
-                                <p>Name: {cert.name}</p>
-                                <p>Course: {cert.course}</p>
-                                <p>Results: {cert.result}</p>
-                              </li>      
-                                          
-                              <li key={key} className="list-group-item py-2">
-                                <p>Issue date: {cert.issueDate}</p>
-                                <p>IC: {cert.ic}</p>
-                              </li>
-                            </ul>
-
-                        </div>
-                    )
-
-                })}
-
-
-           
-
-
-
-
-
+              <CertList
+                certs={this.props.certs} 
+              />
 
               </div>
             </main>
